added authorization so only the owner of the post or the admin can edit/delete it.

+++ BloggerWebApi/Services/PostService.cs
@@ -1,16 +1,19 @@
 using BloggerWebApi.Dto;
 using BloggerWebApi.Interfaces;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Hosting;
 
 namespace BloggerWebApi.Services;
 
 public class PostService : IPostService
 {
     private readonly AppDbContext context;
+    private readonly IHttpContextAccessor httpContextAccessor;
 
-    public PostService(AppDbContext context)
+    public PostService(AppDbContext context, IHttpContextAccessor httpContextAccessor)
     {
         this.context = context;
+        this.httpContextAccessor = httpContextAccessor;
     }
 
     public async Task<IEnumerable<PostPreviewDto>> GetAllAsync()
@@ -34,20 +37,26 @@ public class PostService : IPostService
         return await context.Posts.FindAsync(id);
     }
 
-    public async Task<Post> CreateAsync(Post post)
+    public async Task<Post> CreateAsync(Post post, string userId)
     {
         DateTime date = DateTime.UtcNow;
         post.CreatedDate = date;
         post.LastModifiedDate = date;
+        post.UserId = userId;
         context.Posts.Add(post);
         await context.SaveChangesAsync();
         return post;
     }
 
-    public async Task<Post?> UpdateAsync(int id, Post updatedPost)
+    public async Task<Post?> UpdateAsync(int id, Post updatedPost, string userId)
     {
         var existingPost = await context.Posts.FindAsync(id);
         if (existingPost == null)
+        {
+            return null;
+        }
+
+        if (existingPost.UserId != userId && !httpContextAccessor.HttpContext.User.IsInRole("Admin"))
         {
             return null;
         }
@@ -64,13 +73,20 @@ public class PostService : IPostService
 
 
 
-    public async Task<bool> DeleteAsync(int id)
+    public async Task<bool> DeleteAsync(int id, string userId)
     {
         var post = await context.Posts.FindAsync(id);
         if (post == null)
         {
             return false;
         }
+
+        var currentUser = httpContextAccessor.HttpContext.User;
+        var isAdmin = currentUser.IsInRole("Admin");
+        if (post.UserId != userId && !isAdmin)
+        {
+            return false;
+        }        
 
         context.Posts.Remove(post);
         await context.SaveChangesAsync();
